PTXQC: A package for computing Quality Control (QC) metrics for Proteomics (PTX)

The following sections describe the main components:

Input

Valid input data are either the files from MaxQuant's .txt folder (all versions from MaxQuant >= 1.0 upwards are supported) or a single mzTab file. All mzTab files will work, but most metrics can be obtained from OpenMS' mzTab as produced by the QualityControl TOPP tool (from OpenMS 2.5 onwards).

Important functions

The central function of this package is called createReport and it accepts either MaxQuant or mzTab data, along with a configuration (optional). There is a parser for mzTab MzTabReader and MaxQuant txt files MQDataReader, as well as a plethora of QC metrics derived from a common qcMetric class and scoring functions qual..., e.g. qualGaussDev.

Configuration

The user can modify the behaviour of PTXQC, e.g. to enable/disable certain metrics or change scoring thresholds, via a YAML object/file. By default a Yaml file is written automatically side-by-side to the input files upon running PTXQC for the first time on a particular input. A custom Yaml object can be passed to the main createReport function for customization. Use yaml::yaml.load_file(input = 'myYAML.yaml') to load an existing file and pass the Yaml object along.

Output

Either a PDF and/or Html report which contains QC plots and a description of the metrics.

See also

Author

Maintainer: Chris Bielow chris.bielow@bsc.fu-berlin.de

Other contributors:

Juliane Schmachtenberg [contributor]

Swenja Wagner [contributor]

Patricia Scheil [contributor]

Tom Waschischek [contributor]

Guido Mastrobuoni [data contributor, reviewer]