Boxplots - one for each condition (=column) in a data frame.

Given a data.frame with two/three columns in long format (name, value, [contaminant]; in that order), each group (given from 1st column) is plotted as a bar. Contaminants (if given) are separated and plotted as yellow bars.

boxplotCompare(
  data,
  log2 = TRUE,
  ylab = "intensity",
  mainlab = ylab,
  sublab = "",
  boxes_per_page = 30,
  abline = NA,
  coord_flip = TRUE,
  names = NA
)

Arguments

data

Data frame in long format with numerical expression data

log2

Apply log2 to the data (yes/no)

ylab

Label on Y-axis

mainlab

Main title

sublab

Sub title

boxes_per_page

Maximum number of boxplots per plot. Yields multiple plots if more groups are given.

abline

Draw a horizontal green line at the specified y-position (e.g. to indicate target median values)

coord_flip

Exchange Y and X-axis for better readability

names

An optional data.frame(long=.., short=..), giving a renaming scheme (long->short) for the 'name' column

Value

List of ggplot objects

Details

Boxes are shaded: many NA or Inf lead to more transparency. Allows to easily spot sparse groups