Create a quality control report (in PDF format).

This is the main function of the package and the only thing you need to call directly if you are just interested in getting a QC report.

createReport(
  txt_folder = NULL,
  mztab_file = NULL,
  yaml_obj = list(),
  report_filenames = NULL,
  enable_log = FALSE
)

Arguments

txt_folder

Path to txt output folder of MaxQuant (e.g. "c:/data/Hek293/txt")

mztab_file

Alternative to **txt_folder**, you can provide a single mzTab file which contains PSM, PEP and PRT tables

yaml_obj

A nested list object with configuration parameters for the report. Useful to switch off certain plots or skip entire sections.

report_filenames

Optional list with names (as generated by getReportFilenames). If not provided, will be created internally by calling getReportFilenames.

enable_log

If TRUE all console output (including warnings and errors) is logged to the file given in **report_filenames$log_file**. Note: warnings/errors can only be shown in either the log **or** the console, not both!

Value

List with named filename strings, e.g. $yaml_file, $report_file etc..

Details

You need to provide either a) the folder name of the 'txt' output, as generated by MaxQuant or an mzTab file or b) an mzTab file as generated by the OpenMS QualityControl TOPP tool (other mzTab files will probably not work)

Optionally, provide a YAML configuration object, which allows to (de)activate certain plots and holds other parameters. The yaml_obj is complex and best obtained by running this function once using the default (empty list). A full YAML configuration object will be written in the 'txt' folder you provide and can be loaded using yaml.load.

The PDF and the config file will be stored in the given txt folder.

Note

You need write access to the txt/mzTab folder!

For updates, bug fixes and feedback please visit https://github.com/cbielow/PTXQC.