Generate a Heatmap from a list of QC measurements.

Each list entry is a data.frame with two columns. The first one contains the Raw file name (or the short version). and should be named 'raw.file' (or 'fc.raw.file'). The second column's name must be an expression (see ?plotmath) and contains quality values in the range [0,1]. If values are outside this range, a warning is issued and values are cut to the nearest allowed value (e.g. '1.2' becomes '1'). List entries are merged and columns are ordered by name.

All substrings enclosed by 'X[0-9]*X.' will be removed (can be used for sorting columns). The resulting string is evaluated as an expression. E.g. parse(text = <colname>)

getQCHeatMap(lst_qcMetrics, raw_file_mapping)

Arguments

lst_qcMetrics

List of QCMetric objects

raw_file_mapping

Data.frame with 'from' and 'to' columns for name mapping to unify names from list entries

Value

A ggplot object for printing

Details

To judge the overall quality of each raw file a summary column is added, values being the mean of all other columns per row.